2014/06/16 - Thorsten Some more small fixes in format specifiers of debug output

+++ src/Unix/cygwin/ethernet_cygwin.cpp
@@ -395,7 +395,7 @@ int WinTapEthernetHandler::recv(uint8 *buf, int len)
 	}
 
 	device_total_in += lenin;
-	D(bug("WinTap: Read packet done (len %d)", lenin));
+	D(bug("WinTap: Read packet done (len %d)", (int)lenin));
 	return lenin;
 }
